Improve ToPrimitive
This polyfill's ToPrimitive implementation wasn't according to spec. Fix
it and move it into a separate file so that it can also be imported in
monthcode.ts which can't depend on ecmascript.ts.
